pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "pneumonia"
version = "0.0.1"
description = "Classification task on x-ray images of lungs. The aim is to classify pneumonia tissue."
authors = [
  { name = "Dynamic-Duo-AT", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.128.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "typer==0.15.1",
    "uvicorn==0.40.0",
    "numpy==1.26.4",
    "pillow==10.2.0",
    "matplotlib==3.9.2",
    "wandb>=0.23",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "dvc-gs>=3.0.2",
    "dvc>=3.65.0",
    "dvc-gdrive>=3.0.1",
    "python-multipart>=0.0.21",
    "evidently==0.4.28",
    "scikit-learn>=1.8.0",
    "pandas>=2.3.3",
    "gdown>=5.2.1",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==2.0.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120
lint.select=["E", "F", "I"]

[tool.coverage.run]
omit = ["tests/*"]
